feat: unify demo and sample population across builds
