+++ packages/desktop/src/desktop-feed-media-policy.test.ts
@@ -3,10 +3,10 @@ import { resolve } from "node:path";
 import { describe, expect, it } from "vitest";
 
 describe("desktop feed media policy", () => {
-  it("keeps remote feed media previews reader-only", () => {
+  it("allows inline media only in explicit local feature previews", () => {
     const source = readFileSync(resolve(process.cwd(), "src/App.tsx"), "utf8");
 
-    expect(source).toContain('feedMediaPreviews: "reader-only"');
+    expect(source).toContain('feedMediaPreviews: IS_FEATURE_PREVIEW ? "inline" : "reader-only"');
     expect(source).not.toContain('feedMediaPreviews: "inline"');
   });
 });
