+++ packages/pwa/src/lib/demo-checkpoint.ts
@@ -1,13 +1,5 @@
 import {
-  generateSampleLibraryData,
-  SAMPLE_CHARACTER_ARCS,
-  SAMPLE_CHARACTER_AVATAR_MEDIA,
-  SAMPLE_CURATED_DEMO_MEDIA,
-  sampleCorpusAttribution,
-  sampleCorpusMediaUrl,
-  sampleCorpusSourceUrl,
-  sampleEditorialContentSignals,
-  projectSampleYouTubeVideo,
+  generateDemoLibraryData,
   type Account,
   type FeedItem,
   type Person,
@@ -38,17 +30,6 @@ const DEMO_PUBLIC_KEY = "4".repeat(64);
 const DEMO_CHAIN_DIGEST = "5".repeat(64);
 const DEMO_PAGE_RECORDS = 512;
 export type FreedDemoCheckpointProgressListener = (percent: number) => void;
-const DEMO_CHARACTER_CARE_LEVELS = {
-  "manny-tis": 5,
-  "cygnus-shy": 4,
-  "nudi-branch-manager": 3,
-  "frogbert-angler": 3,
-  "flora-mingo": 2,
-  "nova-remains": 1,
-  "alma-eight": 4,
-  "mora-grey": 2,
-  "colm-still": 1,
-} as const satisfies Readonly<Record<string, Person["careLevel"]>>;
 
 interface FreedDemoCheckpointOptions {
   generatedAt?: number;
@@ -63,247 +44,6 @@ function demoPresentationSeed(): number {
   return values[0]!;
 }
 
-function stablePresentationNumber(value: string, seed: number): number {
-  let hash = seed >>> 0;
-  for (let index = 0; index < value.length; index += 1) {
-    hash = Math.imul(hash ^ value.charCodeAt(index), 16_777_619) >>> 0;
-  }
-  return hash;
-}
-
-function curatedDemoSample(
-  sample: ReturnType<typeof generateSampleLibraryData>,
-  generatedAt: number,
-  presentationSeed: number,
-  previousTopItemId: string | null | undefined,
-) {
-  if (!sample.items.length || !sample.persons.length || !sample.accounts.length || !sample.feeds.length) {
-    throw new Error("The demo requires nonempty record templates.");
-  }
-  const mediaBySha = new Map(SAMPLE_CURATED_DEMO_MEDIA.map((asset) => [asset.sha1, asset]));
-  const mediaByCharacter = new Map<string, typeof SAMPLE_CURATED_DEMO_MEDIA>();
-  for (const arc of SAMPLE_CHARACTER_ARCS) {
-    mediaByCharacter.set(
-      arc.characterId,
-      arc.episodes.flatMap((episode) => {
-        const asset = episode.mediaSha1 ? mediaBySha.get(episode.mediaSha1) : undefined;
-        return asset ? [asset] : [];
-      }),
-    );
-  }
-  const characterItems = new Map<string, FeedItem[]>();
-  let templateIndex = 0;
-  for (const arc of SAMPLE_CHARACTER_ARCS) {
-    const assets = mediaByCharacter.get(arc.characterId) ?? [];
-    const avatar = SAMPLE_CHARACTER_AVATAR_MEDIA.get(arc.characterId) ?? assets[0];
-    const externalId = `${DEMO_BATCH_ID}:sample-character-${arc.characterId}`;
-    const items = arc.episodes.flatMap((episode, sequence) => {
-      // Keep sequence IDs from the authored timeline, including unpublished gaps.
-      if (!episode.mediaSha1) return [];
-      if (!episode.classification) {
-        throw new Error(`Missing editorial classification for ${arc.characterId}:${sequence}`);
-      }
-      const platform = episode.platform ?? arc.platform;
-      const contentType = episode.contentType ?? (
-        platform === "rss" || platform === "medium" || platform === "substack"
-          ? "article" : platform === "youtube" ? "video" : "post"
-      );
-      const asset = episode.mediaSha1 ? mediaBySha.get(episode.mediaSha1) : undefined;
-      if (episode.mediaSha1 && !asset) {
-        throw new Error(`Missing reviewed demo media for ${arc.characterId}:${sequence}`);
-      }
-      if ((platform === "youtube") !== (contentType === "video") ||
-          (episode.video && platform !== "youtube")) {
-        throw new Error("Demo video metadata requires the YouTube video format.");
-      }
-      const video = platform === "youtube"
-        ? projectSampleYouTubeVideo(episode.video, sampleCorpusMediaUrl(asset!), episode.body)
-        : undefined;
-      // The reader displays content text, not link descriptions. Preserve both
-      // independent creators' credits there, including the thumbnail's license.
-      const thumbnailCredit = video && asset
-        ? `Thumbnail by ${asset.creator}, ${asset.license}. Source: ${sampleCorpusSourceUrl(asset)}`
-        : undefined;
-      // Templates supply record defaults, not a ceiling on authored content.
-      // Every public identity and content field is replaced below.
-      const template = sample.items[templateIndex++ % sample.items.length]!;
-      const globalId = `${DEMO_BATCH_ID}:sample-character:${arc.characterId}:${sequence}`;
-      const sourceUrl = video?.sourceUrl ?? (asset
-        ? sampleCorpusSourceUrl(asset)
-        : `https://demo.freed.wtf/?item=${encodeURIComponent(globalId)}`);
-      return [{
-        ...template,
-        globalId,
-        platform,
-        contentType,
-        sourceUrl,
-        author: {
-          id: externalId,
-          displayName: arc.identityNameBase,
-          handle: arc.characterId,
-          avatarUrl: avatar ? sampleCorpusMediaUrl(avatar) : undefined,
-        },
-        content: {
-          text: video ? `${video.text}\n\n${thumbnailCredit}` : episode.body,
-          mediaUrls: asset ? [sampleCorpusMediaUrl(asset)] : [],
-          mediaTypes: asset ? ["image" as const] : [],
-          linkPreview: {
-            url: sourceUrl,
-            title: episode.title,
-            description: video
-              ? `${video.attribution}\n${thumbnailCredit}`
-              : (asset ? sampleCorpusAttribution(asset) : undefined),
-          },
-        },
-        contentSignals: sampleEditorialContentSignals(episode.classification, generatedAt),
-        ...(platform === "rss"
-          ? {
-              rssSource: {
-                feedUrl: `https://sample.freed.wtf/${DEMO_BATCH_ID}/characters/${arc.characterId}`,
-                feedTitle: `${arc.identityNameBase} Field Notes`,
-                siteUrl: "https://sample.freed.wtf",
-              },
-            }
-          : { rssSource: undefined }),
-        ...(arc.location || asset?.coordinates
-          ? {
-              location: {
-                name: arc.location?.name ?? asset!.detail,
-                coordinates: arc.location?.coordinates ?? asset!.coordinates!,
-                source: "text_extraction" as const,
-              },
-            }
-          : { location: undefined }),
-        preservedContent: contentType === "article"
-          ? {
-              preservedAt: generatedAt,
-              publishedAt: generatedAt,
-              author: arc.identityNameBase,
-              text: episode.body,
-              wordCount: episode.body.trim().split(/\s+/).length,
-              readingTime: Math.max(1, Math.ceil(episode.body.trim().split(/\s+/).length / 200)),
-            }
-          : undefined,
-        userState: {
-          ...template.userState,
-          // A replacement demo begins unread, regardless of sample-template history.
-          readAt: undefined,
-          seenSyncedAt: undefined,
-          archived: false,
-          hidden: false,
-        },
-      } satisfies FeedItem];
-    });
-    characterItems.set(arc.characterId, items);
-  }
-
-  const newestFirst: FeedItem[] = [];
-  const maximumEpisodes = Math.max(...[...characterItems.values()].map((items) => items.length));
-  for (let round = 0; round < maximumEpisodes; round += 1) {
-    const activeArcs = SAMPLE_CHARACTER_ARCS
-      .filter((arc) => round < (characterItems.get(arc.characterId)?.length ?? 0))
-      .sort((left, right) =>
-        stablePresentationNumber(`${round}:${left.characterId}`, presentationSeed) -
-        stablePresentationNumber(`${round}:${right.characterId}`, presentationSeed)
-      );
-    for (const arc of activeArcs) {
-      const episodes = characterItems.get(arc.characterId)!;
-      newestFirst.push(episodes[episodes.length - 1 - round]!);
-    }
-  }
-  const visibleTopIndex = newestFirst.findIndex((item) => item.contentType !== "story");
-  const visibleTop = newestFirst[visibleTopIndex];
-  if (visibleTop && visibleTop.globalId === previousTopItemId) {
-    const replacementIndex = newestFirst.findIndex((item, index) =>
-      index !== visibleTopIndex &&
-      item.contentType !== "story" &&
-      item.author.id !== visibleTop.author.id
-    );
-    if (replacementIndex >= 0) {
-      [newestFirst[visibleTopIndex], newestFirst[replacementIndex]] = [
-        newestFirst[replacementIndex]!,
-        newestFirst[visibleTopIndex]!,
-      ];
-    }
-  }
-  const timelineSlots = sample.items
-    .slice(0, newestFirst.length)
-    .map((item) => item.publishedAt)
-    .sort((left, right) => right - left);
-  const items = newestFirst.map((item, index) => {
-    const publishedAt = timelineSlots[index]
-      ?? (timelineSlots.at(-1) ?? generatedAt) - (index - timelineSlots.length + 1) * 60_000;
-    const delta = publishedAt - item.publishedAt;
-    return {
-      ...item,
-      publishedAt,
-      capturedAt: item.capturedAt + delta,
-    };
-  });
-
-  const admittedArcs = SAMPLE_CHARACTER_ARCS.filter((arc) => characterItems.get(arc.characterId)?.length);
-  // Stable membership makes the two feed scopes meaningfully different without
-  // changing who is a friend whenever the timeline is shuffled. Preserve curated
-  // close relationships first, then fill the remaining slots deterministically.
-  const curatedCare = (id: string) =>
-    DEMO_CHARACTER_CARE_LEVELS[id as keyof typeof DEMO_CHARACTER_CARE_LEVELS];
-  const membershipPriority = (id: string) => {
-    const level = curatedCare(id);
-    return level === undefined ? 1 : level >= 3 ? 0 : 2;
-  };
-  const friendIds = new Set([...admittedArcs]
-    .sort((left, right) =>
-      membershipPriority(left.characterId) - membershipPriority(right.characterId)
-      || stablePresentationNumber(left.characterId, 0) - stablePresentationNumber(right.characterId, 0)
-      || left.characterId.localeCompare(right.characterId))
-    .slice(0, Math.round(admittedArcs.length * 0.15))
-    .map((arc) => arc.characterId));
-  const persons = admittedArcs.map((arc, index) => {
-    const template = sample.persons[index % sample.persons.length]!;
-    const assets = mediaByCharacter.get(arc.characterId) ?? [];
-    const isFriend = friendIds.has(arc.characterId);
-    const preferredCare = curatedCare(arc.characterId);
-    // Care level is relationship status, not a second independent rating.
-    const careLevel = (isFriend
-      ? preferredCare !== undefined && preferredCare >= 3 ? preferredCare : 3 + stablePresentationNumber(arc.characterId, 0) % 3
-      : preferredCare !== undefined && preferredCare <= 2 ? preferredCare : 1 + stablePresentationNumber(arc.characterId, 0) % 2
-    ) as Person["careLevel"];
-    return {
-      ...template,
-      id: `${DEMO_BATCH_ID}:sample-person-${arc.characterId}`,
-      name: arc.identityNameBase,
-      bio: arc.bio,
-      avatarUrl: SAMPLE_CHARACTER_AVATAR_MEDIA.get(arc.characterId)?.baseUrl ?? (assets[0] ? sampleCorpusMediaUrl(assets[0]) : undefined),
-      careLevel,
-      relationshipStatus: isFriend ? "friend" : "connection",
-    } satisfies Person;
-  });
-  const accounts = admittedArcs.flatMap((arc, index) => {
-    const template = sample.accounts[index % sample.accounts.length]!;
-    const assets = mediaByCharacter.get(arc.characterId) ?? [];
-    const externalId = `${DEMO_BATCH_ID}:sample-character-${arc.characterId}`;
-    const platforms = [...new Set(characterItems.get(arc.characterId)!.map((item) => item.platform))];
-    return platforms.map((platform) => ({
-      ...template,
-      id: `social:${platform}:${externalId}`,
-      provider: platform,
-      externalId,
-      handle: arc.characterId,
-      displayName: arc.identityNameBase,
-      avatarUrl: SAMPLE_CHARACTER_AVATAR_MEDIA.get(arc.characterId)?.baseUrl ?? (assets[0] ? sampleCorpusMediaUrl(assets[0]) : undefined),
-      personId: persons[index]!.id,
-    } satisfies Account));
-  });
-  const feedTemplate = sample.feeds[0]!;
-  const feeds: RssFeed[] = admittedArcs.filter((arc) => characterItems.get(arc.characterId)!.some((item) => item.platform === "rss")).map((arc) => ({
-    ...feedTemplate,
-    url: `https://sample.freed.wtf/${DEMO_BATCH_ID}/characters/${arc.characterId}`,
-    title: `${arc.identityNameBase} Field Notes`,
-    siteUrl: "https://sample.freed.wtf",
-    imageUrl: undefined,
-  }));
-  return { accounts, feeds, items, persons };
-}
 
 function record(
   registryKey: LibraryCoreCheckpointRegistryKey,
@@ -519,17 +259,9 @@ export function createFreedDemoCheckpointRecords(
   const generatedAt = options.generatedAt ?? Date.now();
   const presentationSeed = options.presentationSeed ?? demoPresentationSeed();
   const previousTopItemId = options.previousTopItemId ?? undefined;
-  const sample = generateSampleLibraryData({
-    batchId: DEMO_BATCH_ID,
-    friendCount: 80,
-    generatedAt,
-    identitiesPerFriend: 2,
-    presentationSeed,
-    previousTopItemId,
-    seed: 20260831,
-    unlinkedIdentityRatio: 1,
+  const curated = generateDemoLibraryData({
+    batchId: DEMO_BATCH_ID, generatedAt, presentationSeed, previousTopItemId,
   });
-  const curated = curatedDemoSample(sample, generatedAt, presentationSeed, previousTopItemId);
   return [
     record("00_checkpoint_header", "checkpoint", {
       authorityEpoch: DEMO_EPOCH_ID,
