+++ packages/pwa/tests/sqlite-opfs-durability.spec.ts
@@ -11,12 +11,12 @@ import {
   type Page,
 } from "@playwright/test";
 import {
-  SAMPLE_SHOWCASE_FEED_COUNT,
-  SAMPLE_SHOWCASE_FRIEND_COUNT,
-  SAMPLE_SHOWCASE_ITEM_COUNT,
-  SAMPLE_SHOWCASE_SOCIAL_IDENTITY_COUNT,
+  generateDemoLibraryData,
 } from "@freed/shared";
 import { pwaOpfsE2eBaseUrl } from "./opfs-e2e-settings";
+const previewPopulation = generateDemoLibraryData({ batchId: "preview-test", generatedAt: 1_788_800_000_000, presentationSeed: 42 });
+const previewCounts = { feeds: previewPopulation.feeds.length, accounts: previewPopulation.accounts.length, items: previewPopulation.items.length, persons: previewPopulation.persons.length };
+
 
 let testOrigin = pwaOpfsE2eBaseUrl;
 let originServer: Server | null = null;
@@ -103,16 +103,19 @@ async function readFacetSummary(page: Page) {
 async function expectShowcaseSampleData(
   page: Page,
   baseline: Awaited<ReturnType<typeof readFacetSummary>>,
+  counts = previewCounts,
 ): Promise<void> {
+  // A reload may still expose the previous complete batch during replacement.
+  await expect(page.getByRole("status").filter({ hasText: /^Loading ·/ })).toHaveCount(0, { timeout: 90_000 });
   await expect
     .poll(() => readFacetSummary(page), { timeout: 90_000 })
     .toMatchObject({
-      rssFeedCount: SAMPLE_SHOWCASE_FEED_COUNT,
-      sampleAccountCount: SAMPLE_SHOWCASE_SOCIAL_IDENTITY_COUNT,
-      sampleFeedCount: SAMPLE_SHOWCASE_FEED_COUNT,
-      sampleItemCount: SAMPLE_SHOWCASE_ITEM_COUNT,
-      samplePersonCount: SAMPLE_SHOWCASE_FRIEND_COUNT,
-      totalCount: baseline.totalCount + SAMPLE_SHOWCASE_ITEM_COUNT,
+      rssFeedCount: counts.feeds,
+      sampleAccountCount: counts.accounts,
+      sampleFeedCount: counts.feeds,
+      sampleItemCount: counts.items,
+      samplePersonCount: counts.persons,
+      totalCount: baseline.totalCount + counts.items,
     });
 }
 
@@ -495,8 +498,8 @@ test("iPhone WebKit persists, clears, and rebuilds the local sample Library", as
     const populated = await readFacetSummary(page);
     const baseline = {
       ...populated,
-      rssFeedCount: populated.rssFeedCount - SAMPLE_SHOWCASE_FEED_COUNT,
-      totalCount: populated.totalCount - SAMPLE_SHOWCASE_ITEM_COUNT,
+      rssFeedCount: populated.rssFeedCount - previewCounts.feeds,
+      totalCount: populated.totalCount - previewCounts.items,
     };
     await openDangerZone(page);
     await expect(
@@ -580,8 +583,8 @@ test("iPhone WebKit completes interrupted sample population after restart", asyn
       opened.page.getByRole("status").filter({ hasText: /^Loading · \d+%$/ }),
     ).toBeVisible({ timeout: 90_000 });
     await expect.poll(() => readFacetSummary(opened.page)).toMatchObject({
-      rssFeedCount: SAMPLE_SHOWCASE_FEED_COUNT,
-      sampleFeedCount: SAMPLE_SHOWCASE_FEED_COUNT,
+      rssFeedCount: previewCounts.feeds,
+      sampleFeedCount: previewCounts.feeds,
       sampleItemCount: 0,
       samplePersonCount: 0,
     });
@@ -757,6 +760,10 @@ test("iPhone WebKit reads pinned reader content after restart without its source
           if (!item) throw new Error("sample reader item is unavailable");
           const readerItem = {
             ...item,
+            globalId: "rss:offline-reader-real-record",
+            contentType: "article" as const,
+            sampleDataFingerprint: undefined,
+            contentSignals: undefined,
             content: {
               ...item.content,
               linkPreview: {
@@ -769,8 +776,11 @@ test("iPhone WebKit reads pinned reader content after restart without its source
             },
             sourceUrl: articleUrl,
           };
+          const store = await import("/src/lib/store.ts");
+          await store.useAppStore.getState().addItems([readerItem]);
+          await runtime.settlePwaLibraryCoreLocalSampleState();
           await readerCache.pinReaderItemInPwa(readerItem);
-          return { globalId: item.globalId };
+          return { globalId: readerItem.globalId };
         } finally {
           await reader.close();
         }
@@ -789,6 +799,7 @@ test("iPhone WebKit reads pinned reader content after restart without its source
     await reopened.goto(`/?item=${encodeURIComponent(pinned.globalId)}`);
     await acceptLegalGate(reopened);
     await waitForLibrary(reopened);
+    await expectShowcaseSampleData(reopened, { totalCount: 1 } as Awaited<ReturnType<typeof readFacetSummary>>);
     const sourceUnavailable = await reopened.evaluate((articleUrl) =>
       fetch(articleUrl)
         .then(() => false)
