+++ packages/shared/src/index.ts
@@ -60,3 +60,5 @@ export * from "./story-wall";
 export * from "./youtube";
 export * from "./sample-youtube";
 export * from "./saved-item-note";
+
+export * from "./demo-data.js";
