+++ packages/ui/src/lib/sample-library-seed.test.ts
@@ -1,4 +1,4 @@
-import { SAMPLE_SHOWCASE_FEED_COUNT, SAMPLE_SHOWCASE_ITEM_COUNT, SAMPLE_SHOWCASE_FRIEND_COUNT, SAMPLE_SHOWCASE_SOCIAL_IDENTITY_COUNT } from "@freed/shared";
+import { generateDemoLibraryData, DEMO_POPULATION_COUNTS } from "@freed/shared";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import {
   clearSampleLibraryDataWithProgressToast,
@@ -23,8 +23,8 @@ describe("sample Library seeding", () => {
     const initialize = vi.fn(async () => undefined);
     const onProgress = vi.fn();
     const addSampleLibraryData = vi.fn(async (data, listener) => {
-      expect(data.feeds).toHaveLength(SAMPLE_SHOWCASE_FEED_COUNT);
-      expect(data.items).toHaveLength(SAMPLE_SHOWCASE_ITEM_COUNT);
+      expect(data.feeds).toHaveLength(DEMO_POPULATION_COUNTS.feeds);
+      expect(data.items).toHaveLength(DEMO_POPULATION_COUNTS.items);
       listener?.({ percent: 40, phase: "items" });
     });
     const seedSocialConnections = vi.fn();
@@ -47,6 +47,27 @@ describe("sample Library seeding", () => {
       phase: "items",
     });
     expect(seedSocialConnections).toHaveBeenCalledOnce();
+  });
+
+  it("uses the full curated demo with a fresh seed for every sample population", async () => {
+    const random = vi.spyOn(globalThis.crypto, "getRandomValues")
+      .mockImplementationOnce(array => { (array as Uint32Array)[0] = 42; return array; })
+      .mockImplementationOnce(array => { (array as Uint32Array)[0] = 123; return array; });
+    const addSampleLibraryData = vi.fn(async () => undefined);
+    const actions = { initialize: vi.fn(async () => undefined), isInitialized: true, addSampleLibraryData };
+    try {
+      await refreshSampleLibraryData(actions);
+      await refreshSampleLibraryData(actions);
+      const batches = addSampleLibraryData.mock.calls as unknown as [ReturnType<typeof generateDemoLibraryData>][];
+      const first = batches[0]![0];
+      const second = batches[1]![0];
+      expect(first.items).toHaveLength(DEMO_POPULATION_COUNTS.items);
+      expect(first.feeds.every(feed => !feed.enabled)).toBe(true);
+      expect(first.persons.filter(person => person.relationshipStatus === "friend").length).toBe(Math.round(first.persons.length * 0.15));
+      expect(first.items.every(item => item.content.mediaUrls?.[0]?.startsWith("https://"))).toBe(true);
+      expect(first.items.map(item => item.content.text)).not.toEqual(second.items.map(item => item.content.text));
+      expect(first.items[0]!.sampleDataFingerprint?.batchId).not.toBe(second.items[0]!.sampleDataFingerprint?.batchId);
+    } finally { random.mockRestore(); }
   });
 
   it("formats a locale-aware progress label", () => {
@@ -76,7 +97,7 @@ describe("sample Library seeding", () => {
     expect(useToastStore.getState().toasts).toHaveLength(1);
     expect(useToastStore.getState().toasts[0]).toMatchObject({
       message:
-        `Sample data added: 100%. ${SAMPLE_SHOWCASE_FEED_COUNT.toLocaleString()} feeds, ${SAMPLE_SHOWCASE_ITEM_COUNT.toLocaleString()} items, ${SAMPLE_SHOWCASE_FRIEND_COUNT.toLocaleString()} friends, and ${SAMPLE_SHOWCASE_SOCIAL_IDENTITY_COUNT.toLocaleString()} social identities.`,
+        `Sample data added: 100%. ${DEMO_POPULATION_COUNTS.feeds.toLocaleString()} feeds, ${DEMO_POPULATION_COUNTS.items.toLocaleString()} items, ${DEMO_POPULATION_COUNTS.persons.toLocaleString()} people, and ${DEMO_POPULATION_COUNTS.accounts.toLocaleString()} social identities.`,
       type: "success",
     });
   });
